Add files via upload
Added calculations for gauges that require some math and made separate gauges for those to be used in the JS file.
Also made less of a delay in loading the information and having it ready to be accessed by the user. One file (USWaterlog2.txt) is the working version that is written to as the data is downloaded. After all that data is downloaded, it overwrites the second file (USWaterlog.txt) and instantly becomes available.