Add files via upload
Adjusted code to work with new gauges that were created in the PHP file. These new gauges account for the math that was done in JS before.

Also adjust visual sections to use the flow that is in index 1 of their flow array. This doesn't seem to be working smoothly, so I will address.