Add files via upload
updated rives.json from extract.js to not include extra text in the info content of visual river sections. 
Also updated to include the new virtual gauges.
usingJSON updated to put this information in and include the related flow with visual river sections.
Looks almost good to go as the main page now.