Add files via upload
Data files that identify the USGS gauges that do not have any predictions. We'll try to use some modeling for those missing gauges.

Slight change to USGS_NOAA_corr to make this data possible. 

In the future: we will be predicting on the USGS_missing using the NOAA list of gauges.